Re-introduced config.php and doc updates
includes/core.php now includes includes/config.php and in this way
the php files are considered functionaly seperate. core.php defines
functions and config.php defines deployment-specific paths and
database/table names.

Additionally updated documentation. It's not always intuitive how
to set up an http server to serve something like this so I may
write INSTALL.arch and INSTALL.gentoo files to keep a record of the
install process on both operating systems. Mainly the gentoo docs
will have the necessary APACHE2_MODULES line for definition in
make.conf, because this upset me for a while.

generate_passwd.php is included in this commit also to aid system
administrators in testing the build, as this page provides you with
a hashed password in the SHA512-CRYPT format with a random salt which
can be set as a user's password in the sql database. Again, this is
only for testing.

It is not my intention to ever create an online account creation
interface. If I ever do, it will be secured behind a one-time key
generated by the systems administrator. bmffd was not developed
to be an open file-sharing service, but only functions as a
secure, end-user friendly (meaning that windows friends can access
this w/o downloading anything extra) solution to file-sharing among
a hardly-expanding.