Added the message pile for reporting issues
Message pile is intended for end-users to report bugs to me, as I
use this in a deployment environment. This may or may not be useful
to you; if it is not, then feel free to delete the message_pile directory
and never create the message_pile SQL table as detailed in the INSTALL

Miscellaneous bug fixes are included in this build as well, some things
related to change_password I think.

Also improved the logout button