One-time account creation and user listing
You can now run the ./otp utility to generate a list of one-time
passwords, one of which can be distributed to create a single account

Works by challenging the user to submit the previous entry in the hashchain.
So if you have just generated the hashchain, the PHP page looks for a user
to submit the correct SHA-1 hash which encrypts to the current challenge hash,
so essentially the server says "Give me hash #99 so I can check if that hashes to
form hash #100, which I have stored." If it is successful, the user account is created,
and the server store the user-entered hash (hash #99) and waits for the next
user to enter hash #98, which it can then verify using the same method.

Oh, and I added a user-listing page. Currently this relies on a `domain`
column, as I have an email service running on the same server. If you have
a problem simply delete the mailto: links in user.php