bmfft has been added!
bmfft (big mike's fancy file tagger) will be everything I've ever dreamed of:
it indexes files on the localhost with a list of tags and other metadata,
which is then searchable through the web interface.

this is planned to splinter off into a different project and replace
the aging gallery soon, but not immediately.

to use: run ./bmfft (directory to index) (tag file to create)

to generate the tag file. Then the web interface will allow one to
browse, tag, and search for files in this database. Anytime you add
files to the indexed folder, make sure to run the above command again
(this is non-destructive, as each file is stored according to its
md5 hash). . . neat!

lots more fixes and features to come~