DB indexing updates, popular page for easy tag browsing
Search results are now paged, because it eats a lot of bandwidth
to load 100 results at once. Need to add thumbnailing soon for
this reason as well.

I'll be taking a break from a lot commits as I tag my files and do
everything else that I've been wanting to do for a while now~