Tagging frame added
When viewing a file, a list of tags is presented in a new frame
where the mascor typically is, allowing the user to edit tags
without clicking on the picture to open a new window.

This improves the speed of hooYa! as opening and closing windows is
evidently a resource-consuming process, and makes hooYa! feel more
unified in style. Also it eliminates that nasty flickering when opening
new tabs, which I quickly got tired of