MicroTXT - Complaints general
Integrated MicroTXT (https://github.com/beardog108/MicroTXT) as a local
textboard for only authenticated users to post in. Mostly for bugs and
fun but if you don't have any other means of communication this could be
a cutesy little forum

Would have put this in as a submodule but since I'm reclaiming a lot
of the source and changing it around I've just included it here.