GNU Database Manager -> MySQL transition
Massive transition from a static DB file to a MySQL database model.
Much faster and fuller data access, given MySQL's keyed collections
and ability to search through a database.

Search does not function currently, but the random browse provides
a good example of the direction and agility of the new site. Not
many UI changes, except that tags and namespaces have been merged
into `Tags` which are just a Space->Member mapping. So something
like 'place'->'beach' is an appropriate tag