Underscore replacement provided for search bar
Because tags are always in the form of spaces_replaced_by_underscores
for existing sites like *booru, I have decided to expect provided
tags to be in this form. To accomplish this, when the user presses
space as he would when typing "akemi homura" s/he will now find
that the input field reads "akemi_homura". This is much shorter and
faster than entering an underscore by hand.

An underscore may be "escaped" by entering two spaces. This is how
search terms are seperated. So if you wanted to search for all pictures
of homura akemi w/ a gun, your search query may look like:

akemi_homura gun