tagspace:member searching, like 'series:pmmm'
After the SQL update, this feature was not reimplemented. Consider this
the full reimplementation. This makes it far easier to find rogue tagspaces
like if someone mispells 'series' as 'sereis', you can do a quick clickthrough
the popular/ tags and narrow down the picture, then delete that tagspace:member pair