Users may set properties which are not immutable
If the property is immutable (set in includes/database.php) then
an input field is printed and the user is allowed to change that
property.

Otherwise, the property value is printed and all attempts to set
the property value for the immutable property are disregarded.

Immutable properties are not recognized outside of the PHP environment
This means that you may set the image height to anything you care to
from the SQL environment; this is obviously not recommended

In short, immutability is strictly enforced on the front-facing parts
of hooYa!