Pass multiple --property parametes, either as a literal or a regular expression
The script is nao smart enough to detect if you mean to give it a
regex or a literal for a --property flag. If you have some enclosing
parentheses in your statement of the property format, the format
is interpreted as a regex; otherwise it is taken literally

For example, I just ran a command like

./divinity --import --directory (DIR) --class anime
--property 'season:1' --property 'episode:- (\d+)'

This scans (DIR) recursively and maps all files in hooYa as is
natural. The important part is the --property flags. 'season:1' is
interpreted literally to be season->1, while 'episode:- (\d+)' is
interpreted to be a regex, and the filename is matched against this.

The parentheses indicate which part of the filename is to be captured
as the episode number