Tag aliases
A SQL table has been set up to map aliases -> actual tag names. This
will help with things like pmmm -> puella_magi_madoka_magica. It would
be useful to list all of the aliases that a user can use, but this
is out-of-scope for nao

Additionally there deserves to be an area for users to add their own
alises but this is also OOS