divinity: --absorb to map files into hooYa cold storage
For your massive art collections, using something like

./divinity --import --absorb --directory=/var/media/pictures/

moves all the files in /var/media/pictures/ into hooYa cold storage
as defined by the PHP script in includes/config.php
"CONFIG_HOOYA_STORAGE_PATH". Of course, they are additionally mapped
into the database and able to be tagged; the only difference is that
the original file is MOVED and put into the storage directory.

It's not recommended to import large files like movies
and stuff like this because they filename is clobbered when moved
into storage