New Layout
The newly-created Recent Activity page, which shows all pictures
which have gotten a new tag recently, replaces the old index.php
page as the new front-page for the site.

I hope the anime girls on the front-page make the project's implications
more resonant.

Otherwise, the old search page has been moved to power.php because
the interface is overwhelming for low-power users.

There's also a cute search bar at the left-hand side of a lot of pages!